changed command line interface of python running scripts: 	-i, --inputDat: with name of configuration .dat 	--dryRun      :	configure LoopAll but dont actually run it (check all is ok!) 	-n, --nJobs   : number of jobs (reduce) 	-j, --jobId   : job ID (reduce)

+++ PhotonAnalysis_scripts/fitter.py
@@ -5,6 +5,8 @@ from sys import argv
 from python.configProducer import *
 from python.lumi import *
 import sys
+from python.runOptions import *
+(options,args)=parser.parse_args()
 
 ROOT.gSystem.Load("libRooFit.so")
 ROOT.gSystem.Load("libPhysics.so");
@@ -16,45 +18,36 @@ ROOT.gROOT.SetBatch(1)
 ROOT.gBenchmark.Start("Analysis");
 
 config_file="datafiles.dat"
-if len(sys.argv) > 1:
-    config_file = sys.argv[1]
-
-njobs=-1
-jobId=0
-if len(argv) > 2:
-    if len(argv) == 3:
-        print "Usage: fitter.py <config_file> [<njobs> <job_id>]" 
-        exit(1)
-    njobs = int(argv[2])
-    jobId = int(argv[3])
+if options.inputDat:
+    config_file = options.inputDat
 
 ut = ROOT.LoopAll();
-cfg = configProducer(ut,config_file,0,njobs,jobId)
-  
-ut.LoopAndFillHistos();
-ROOT.gBenchmark.Show("Analysis");
+cfg = configProducer(ut,config_file,0,int(options.nJobs),int(options.jobId))
 
-ut.WriteFits();  
+if not options.dryRun:
+  ut.LoopAndFillHistos();
+  ROOT.gBenchmark.Show("Analysis");
 
-jsonname = "%s.json" % ( ut.histFileName.rsplit(".root")[0] )
+  ut.WriteFits();  
 
-if "/" in str(ut.histFileName):
+  jsonname = "%s.json" % ( ut.histFileName.rsplit(".root")[0] )
+
+  if "/" in str(ut.histFileName):
     path,name = str(ut.histFileName).rsplit("/",1)
     ut.histFileName=path+"/histograms_"+name
-else:
+  else:
     ut.histFileName="histograms_"+ut.histFileName
-ut.WriteHist();  
+  ut.WriteHist();  
 
-print "Producing JSON file for lumi calculation.."
-runLines = dumpLumi([ut.histFileName])
+  print "Producing JSON file for lumi calculation.."
+  runLines = dumpLumi([ut.histFileName])
 
-print "Processed lumi sections: "
-print "{" + ", ".join(runLines) + "}"
+  print "Processed lumi sections: "
+  print "{" + ", ".join(runLines) + "}"
 
-json = open( "%s" % jsonname, "w+" )
-print >>json, "{" + ", ".join(runLines) + "}"
-json.close()
+  json = open( "%s" % jsonname, "w+" )
+  print >>json, "{" + ", ".join(runLines) + "}"
+  json.close()
 
-ut.WriteCounters();  
-
+  ut.WriteCounters();  
 ROOT.gROOT.Reset()
